Create blockchain.py
Partnered with 6point6 to complete a group project with vision to produce a strategic proposal of security on how security operations could be revolutionized by use of blockchain.
Formulated innovation strategies and analyzed opportunities for small scale and large scale application of concept. Project was awarded as best among other group projects completed with organization.